+++ app/controllers/ItemRecord.scala
@@ -18,7 +18,7 @@ object ItemRecord extends REST {
 
   protected def single(row: Row): JsValue = {
     row match {
-      case Row(id: Long, number: Double, item_id: Int, stat_record_id: Int) => {
+      case Row(id: Long, number: Float, item_id: Long, stat_record_id: Long) => {
         JsObject(
           "id"             -> JsNumber(id) ::
           "number"         -> JsNumber(number) ::
